requirements.txt
Flask
Flask-JWT-Extended
Flask-CORS
python-dotenv
bcrypt
supabase
requests
gunicorn
